Add Pause Menu flow
