+++ Assets/_Project/Scenes/PopUps/PauseMenu.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: bf67b2822078ff64e82935ec8637f1ea
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
