+++ Assets/_Project/Screen_PauseMenu.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 68056d2fe8bf15542aafb8eab9729a0f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
