+++ Assets/_Project/Screen_PauseMenu/PauseMenuRef.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6b5295a8ec073ac469abf81890e3356a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
