+++ Assets/_Project/Source/SettingsMenuController.cs
@@ -8,8 +8,6 @@ namespace Source
     public class SettingsMenuController : BaseScreen
     {
         [SerializeField] private Button _closeButton;
-        [SerializeField] private Button _mainMenuButton;
-        [SerializeField] private ScreenReference _mainMenuScreenRef;
         [FoldoutGroup("Settings UiOverlay elements")]
         [SerializeField] private Slider _sliderMaster;
         [FoldoutGroup("Settings UiOverlay elements")]
@@ -33,7 +31,6 @@ namespace Source
         {
             base.Initialize();
             _closeButton.onClick.AddListener(CloseButtonClickHandler);
-            _mainMenuButton.onClick.AddListener(MainMenuButtonClickHandler);
 
             _sliderMaster.onValueChanged.AddListener(OnMasterVolumeChangeHandler);
             _sliderMusic.onValueChanged.AddListener(OnMusicVolumeChangeHandler);
@@ -41,11 +38,6 @@ namespace Source
             _sliderUiSfx.onValueChanged.AddListener(OnUiSfxVolumeChangeHandler);
 
             SyncSlidersFromMixers();
-        }
-
-        private void MainMenuButtonClickHandler()
-        {
-            ScreenService.LoadSingleSceneAsync(_mainMenuScreenRef);
         }
 
         private void OnMasterVolumeChangeHandler(float value)
