+++ ProjectSettings/EditorBuildSettings.asset
@@ -21,6 +21,9 @@ EditorBuildSettings:
     path: Assets/_Project/Scenes/PopUps/GamePlayUi.unity
     guid: faf56d28d9362c5478a7bda7c32fc4a7
   - enabled: 1
+    path: Assets/_Project/Scenes/PopUps/PauseMenu.unity
+    guid: bf67b2822078ff64e82935ec8637f1ea
+  - enabled: 1
     path: Assets/_Project/Scenes/LevelSelect.unity
     guid: 4736d2832026cc44aa35fa6de3fee5b3
   - enabled: 1
